pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "print-ai-monitor"
version = "0.1.0"
description = "Local OctoEverywhere webhook service that powers off a TAPO plug on AI print failure."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "python-dotenv>=1.0,<2.0",
  "python-kasa>=0.10,<1.0",
  "uvicorn>=0.34,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1.0",
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.25,<1.0",
]

[project.scripts]
print-ai-monitor = "print_ai_monitor.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
